+++ test/fixtures/provide-plugin-custom/webpack.config.js
@@ -1,0 +1,15 @@
+'use strict';
+
+module.exports = {
+  mode: 'development',
+  context: __dirname,
+  stats: 'none',
+  entry: './foo.js',
+  output: {
+    path: '/',
+  },
+  node: false,
+  infrastructureLogging: {
+    level: 'warn',
+  },
+};
